requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.1
PyJWT==2.8.0
boto3==1.34.0
numpy==1.26.2
gunicorn==21.2.0
weasyprint==60.0
matplotlib==3.8.0 